+++ packages/kernel/src/host/model/index.ts
@@ -1,10 +1,8 @@
 import {Client} from '@otakustay/ipc';
 import {Protocol} from '@oniichan/host/server';
 import {newUuid} from '@oniichan/shared/id';
-import {ChatMessagePayload, createModelClient, ModelClient} from '@oniichan/shared/model';
+import {ChatMessagePayload, createModelClient, ModelClient, isModelConfigValid} from '@oniichan/shared/model';
 import {ModelUsageTelemetry} from '@oniichan/storage/telemetry';
-// TODO: Move this to shared
-import {isModelConfigValid} from '@oniichan/host/utils/config';
 import {CodeResult, streamingExtractCode} from './extract';
 
 export class ModelAccessHost {