+++ packages/web/src/server/handlers/compose.ts
@@ -1,5 +1,5 @@
 import {getDefaultStore} from 'jotai';
-import {clientAtom} from '@/atoms/client';
+import {ipcAtom} from '@/atoms/ipc';
 import {editingAtom} from '@/atoms/draft';
 import {RequestHandler} from '@otakustay/ipc';
 
@@ -9,7 +9,7 @@ export class ComposeNewMessageRequestHandler extends RequestHandler<void, void> 
     // eslint-disable-next-line require-yield
     async *handleRequest(): AsyncIterable<void> {
         const store = getDefaultStore();
-        await store.get(clientAtom);
+        await store.get(ipcAtom);
         store.set(editingAtom, {threadUuid: crypto.randomUUID(), mode: 'new'});
     }
 }